+++ packages/enzyme-utilities/package.json
@@ -23,10 +23,12 @@
   },
   "homepage": "https://github.com/Shopify/quilt/blob/master/packages/enzyme-utilities/README.md",
   "dependencies": {
-    "@shopify/javascript-utilities": "^2.1.0",
     "lodash": "^4.17.5"
   },
   "devDependencies": {
+    "@shopify/javascript-utilities": "^2.1.0",
+    "enzyme": "^3.3.0",
+    "react": "^16.3.2",
     "typescript": "~2.7.2"
   }
 }
